==> R/01_extract_modelled_data_SO.Rmd <==
---
title: "Extract data from ecosystem models for the Southern Ocean"
author: "Camilla Novaglio"
date: "27/04/2023"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

### AIM

Extract mean size-resolved total consumer biomass across FishMIP models 

### Data source 

Size-resolved biomass (tcblog10) is as for EmergentConstrint project. Adding macroecological for this project: 

Download on 01/05/2023 - only history (and picontrol as in same folder) and not future as this project does not require it

scp -r b381217@levante.dkrz.de:/work/bb0820/ISIMIP/ISIMIP3b/OutputData/marine-fishery_global/MACROECOLOGICAL/gfdl-esm4/historical/*nat_default_tcblog10* ./ 

scp -r b381217@levante.dkrz.de:/work/bb0820/ISIMIP/ISIMIP3b/OutputData/marine-fishery_global/MACROECOLOGICAL/ipsl-cm6a-lr/historical/*nat_default_tcblog10* ./ 

### SO boundaries 

Latitudinal limits (-40) to mimic Yang et al. 2022

### Checked 

Final values are mean g m-2 for the period 1990-1999 -> for models providing monthly values these are 1) annual means and 2) period means; for models providing annual values these are period means. These outputs have no time dimension (i.e. they represent a stock and not a rate, as opposed to NPP which is in s-1 or to catches which should be in e.g. year-1). Hence, here to get annual values we average instead of sum across months.       

To transform from g m-2 to g C m-2 for better comparison with inputs I did /10 as suggested by Ryan, common assumption. 
Also refs: 
https://aslopubs.onlinelibrary.wiley.com/doi/full/10.1002/lno.11709 https://link.springer.com/article/10.1007/s00227-011-1783-7
how much of the fish body is carbon? 10-15%

MACROECOLOGICAL: 5 bins, size classes 1 to 5 - checked with Ryan. 

### Set environment 

```{r environment, echo = FALSE}

rm(list=ls())

library(ncdf4) 
library(raster) 
library(ggplot2) 
library(data.table)
library(dtplyr)
library(ncdf4.helpers)
library(tictoc)
library(dplyr, warn.conflicts = FALSE)
library(broom)
library(tibble)
library(tidyr)
library(purrr)
library(stringr)
library(parallel)
library(pbapply)

select<-dplyr::select
summarise<-dplyr::summarise

# load functions 
source("/home/ubuntu/Extract_SouthernOcean/R/Helper_functions_SO.r")

```

### Read in files and extract modelled data at global scale

```{r step 1 to 2}

# create file names ----
dir<-"/rd/gem/private/users/camillan/EmergentConstraintData" 

mem<-c("apecosm","boats", "dbpm", "zoomss", "ecotroph", "macroecological") 
esm<-c("gfdl-esm4", "ipsl-cm6a-lr")
scenario<-c("historical") 

# all 
combinations<-expand.grid(mem = mem, esm = esm, scenario = scenario) %>%
  mutate(resolution = ifelse(mem %in% c("zoomss", "ecotroph", "macroecological"), "annual", "monthly"), 
         year = case_when(
           mem %in% c("boats", "zoomss", "ecotroph", "macroecological") & scenario == "historical" ~ "1950_2014", 
           mem %in% c("apecosm", "dbpm") & scenario == "historical" ~ "1850_2014",
           scenario %in% c("ssp126", "ssp585") ~ "2015_2100"), 
         mem = as.character(mem), 
         esm = as.character(esm), 
         scenario = as.character(scenario),
         netcdf_name = paste0(paste(mem, esm, "nobasd", scenario, "nat_default_tcblog10_global", resolution, year, sep ="_"), ".nc"),
         esm_simpler = ifelse(esm == "ipsl-cm6a-lr", "ipsl","gfdl"),
         identifier = paste(mem, esm_simpler, scenario, sep ="_")) %>% 
  select(-esm_simpler) %>% 
  arrange(mem, esm, scenario)

# apply function in // ----

netcdf = combinations$netcdf_name

history_year1 = 1995 #1990 (see comments on 02 for reference period)
history_year2 = 2014 #1999
future_year1 = NA
future_year2 = NA

brick_data_annual_subset_mean_ant<-mclapply(netcdf, function(x) extract_antarctica(x, file = "new", history_year1=history_year1, history_year2=history_year2, future_year1=future_year1, future_year2=future_year2), mc.cores = detectCores()-2)

# mclapply returns results in the correct order: https://stackoverflow.com/questions/14697901/is-mclapply-guaranteed-to-return-its-results-in-order#:~:text=Short%20answer%3A%20it%20does%20return,an%20R%20function...)
names(brick_data_annual_subset_mean_ant) <- combinations$identifier

# NOTE
# land-sea mask between GFDL and IPSL is different so coastal values might be "biased" e.g. mean with na.rm=TRUE will consider only GFDL values for the very coastal gridcell (and only a set of models). use na.tm=FALSE so that these cells are NAs

```

### multimodel means 

```{r}

# do size classes 2:5 that are in common to all models 
removeNA = "narmF" # narmT or narmF

mean_2_5 <-list()

for(i in 2:5){ 

  trial<-stack(
    brick_data_annual_subset_mean_ant$apecosm_ipsl_historical[[i]],
    brick_data_annual_subset_mean_ant$boats_gfdl_historical[[i]], 
    brick_data_annual_subset_mean_ant$boats_ipsl_historical[[i]],
    brick_data_annual_subset_mean_ant$dbpm_ipsl_historical[[i]], 
    brick_data_annual_subset_mean_ant$ecotroph_gfdl_historical[[i]],
    brick_data_annual_subset_mean_ant$ecotroph_ipsl_historical[[i]],
    brick_data_annual_subset_mean_ant$macroecological_gfdl_historical[[i]],
    brick_data_annual_subset_mean_ant$macroecological_ipsl_historical[[i]],
    brick_data_annual_subset_mean_ant$zoomss_gfdl_historical[[i]], 
    brick_data_annual_subset_mean_ant$zoomss_ipsl_historical[[i]])
 
  if(removeNA == "narmT"){
    mean_2_5[[i]] <- stackApply(trial, indices =  rep(1,nlayers(trial)), fun = mean, na.rm = TRUE)
  }else if (removeNA == "narmF"){
    mean_2_5[[i]] <- stackApply(trial, indices =  rep(1,nlayers(trial)), fun = mean, na.rm = FALSE)
  } 
}

mean_2_5[sapply(mean_2_5, is.null)] <- NULL

# do size class 1 where macroecological is available but not boats

trial_1<-stack(
    brick_data_annual_subset_mean_ant$apecosm_ipsl_historical[[1]],
    brick_data_annual_subset_mean_ant$dbpm_ipsl_historical[[1]],
    brick_data_annual_subset_mean_ant$ecotroph_gfdl_historical[[1]],
    brick_data_annual_subset_mean_ant$ecotroph_ipsl_historical[[1]],
    brick_data_annual_subset_mean_ant$macroecological_gfdl_historical[[1]],
    brick_data_annual_subset_mean_ant$macroecological_ipsl_historical[[1]],
    brick_data_annual_subset_mean_ant$zoomss_gfdl_historical[[1]],
    brick_data_annual_subset_mean_ant$zoomss_ipsl_historical[[1]])


if(removeNA == "narmT"){
  mean_1<-stackApply(trial_1, indices =  rep(1,nlayers(trial_1)), fun = mean, na.rm = TRUE)
}else if (removeNA == "narmF"){
  mean_1<-stackApply(trial_1, indices =  rep(1,nlayers(trial_1)), fun = mean, na.rm = FALSE)
}

# do size class 6 where boats and macroecological are not available

trial_6<-stack(
    brick_data_annual_subset_mean_ant$apecosm_ipsl_historical[[6]],
    brick_data_annual_subset_mean_ant$dbpm_ipsl_historical[[6]], 
    brick_data_annual_subset_mean_ant$ecotroph_gfdl_historical[[6]],
    brick_data_annual_subset_mean_ant$ecotroph_ipsl_historical[[6]],
    brick_data_annual_subset_mean_ant$zoomss_gfdl_historical[[6]], 
    brick_data_annual_subset_mean_ant$zoomss_ipsl_historical[[6]])

if(removeNA == "narmT"){
  mean_6<-stackApply(trial_6, indices =  rep(1,nlayers(trial_6)), fun = mean, na.rm = TRUE)
}else if(removeNA == "narmF"){
  mean_6<-stackApply(trial_6, indices =  rep(1,nlayers(trial_6)), fun = mean, na.rm = FALSE)
}

# append all data together 
# append(myList, item, after = index)
mean_all<-append(mean_2_5, mean_1, after = 0)
mean_all<-append(mean_all, mean_6, after = 5)

# # CHECK - OK
# # calculate mean using other approach for size class 2 and compare results to above approach
# # NAs are removed by def when doing rasterToPoints() so this comparison works only when removeNA == "narmT"
# 
# trial<-lapply(brick_data_annual_subset_mean_ant, "[[", 2)
# trial[sapply(trial, is.null)] <- NULL
# 
# trial2<-lapply(trial, function(x) as.data.frame(rasterToPoints(x)))
# trial2$model = gsub("[^a-zA-Z]", "", rownames(trial2))
# 
# trial3<-do.call(rbind.data.frame, trial2) %>%
#   group_by(x,y) %>%
#   summarise(mean = mean(index_1, na.rm=removeNA)) %>% 
#   ungroup()
# 
# trial_raster<-rasterFromXYZ(trial3)
# 
# # same outputs
# plot(trial_raster)
# plot(mean_all[[2]])

```

## save means 

```{r}

save(mean_all, file = paste("/home/ubuntu/Extract_SouthernOcean/Tmp_data/",  Sys.Date(), "_01_extract_modelled_data_SO_means_", removeNA, "_Hyear_",history_year1, ".RData", sep = ""))

# # old outputs have this format.  
# 
# if(removeNA == "narmT"){
#   save(mean_all, file = "/home/ubuntu/Extract_SouthernOcean/Tmp_data/01_extract_modelled_data_SO_means.RData")}else if(removeNA == "narmF"){
#   save(mean_all, file = "/home/ubuntu/Extract_SouthernOcean/Tmp_data/01_extract_modelled_data_SO_means_narmF.RData")
#     }

```
