==> R/03_work_with_empirical.Rmd <==
---
title: "Explore and consider empirical data from Yang et al. 2022"
author: "Camilla Novaglio"
date: "22/05/2023"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

### AIM

consider phyto and zooplankton data fro Yang et al. 2022

### Data source 

Provided by Svenja (Yang email with data attached and explained) and moved to the project Tmp_data folder: 
Zooplankton_summer_data.csv
chla_summer_data_Dec_Mar.csv

### Set environment 

```{r environment, echo = FALSE}

rm(list=ls())

library(raster) 
library(ggplot2) 
library(dplyr)
library(sf)
library(rnaturalearth)
library(data.table)

select<-dplyr::select
summarise<-dplyr::summarise

```

## Load adn explore data  

```{r}

zoop<-fread("/home/ubuntu/Extract_SouthernOcean/Input/Zooplankton_summer_data.csv", data.table = FALSE)
# head(zoop)
# from Email Svenja 
# you can find the "summer data" (n=2028) in sheet 2 of the attachment. The mean biomass in each cell (2° latitude × 5° logitiude) is just the simple arithmetic mean of all biomass in that cell. For example, if 40 records were found in a specific cell (160°E-165°E, 60°S-62°S), the value shown in our Fig. 2b is the mean of these 40 value. You can use "fishnet" tool in Arcgis (Arctoolbox/Data Management Tools/Sampling/Create Fishnet) to finish the average procedure. I wish this helps and feel free to contact me if you have any questions about the data or the use of Arcgis.

# approach 1 - first raster then average 
# https://gis.stackexchange.com/questions/79062/how-to-make-raster-from-irregular-point-data-without-interpolation
colnames(zoop) <- c('X', 'Y', 'Z')
zoop<-zoop %>% filter(Y <= -40) # consider -40 only as per plankton and modeleld data
zoop<-as.matrix(zoop)

e <- extent(zoop[,1:2])
# e <- extent(-180, 180, -90, -40) # change to match modelled data
# # this needs to be 2 lat by 5 long 
columns<-360/5 
rows<-(abs(e[3])-abs(e[4]))/2 # 38/2
r <- raster(e, ncol=columns, nrow=rows) 
x <- rasterize(zoop[, 1:2], r, zoop[,3], fun=mean)
plot(x)

zoop_raster<-x
crs(zoop_raster) <- "+proj=longlat +datum=WGS84 +no_defs" 

phyt<-fread("/home/ubuntu/Extract_SouthernOcean/Input/chla_summer_data_Dec_Mar.csv", data.table = FALSE)
# from Email Svenja 
# Please find attached the Chl-a data from Guang to plot the circumpolar phytoplankton data. If I understood correctly, the data are in Chl mg m-3. To convert to depth-integrated g Chl m-2, you need to multiply the values by 41. Then, they assumed a C:Chl-a ratio of 60.5.

colnames(phyt) <- c('X', 'Y', 'Z')

# checked with Svenja
phyt$Z<-phyt$Z*41
phyt$Z<-phyt$Z*60.5
phyt$Z<-phyt$Z/1000 # from mg to g

max(phyt$X)
min(phyt$X)
max(phyt$Y)
min(phyt$Y)

phyt<-as.matrix(phyt)
e <- extent(phyt[,1:2])
# e <- extent(-180, 180, -90, -40) # change to match modelled data
columns<-360/5 
rows<-(abs(e[3])-abs(e[4]))/2 # same as above
r <- raster(e, ncol=columns, nrow=rows) 
x <- rasterize(phyt[, 1:2], r, phyt[,3], fun=mean)
plot(x)

# fix map 
bb <- extent(-180, 180, e[3], e[4])
x <- setExtent(x, bb, keepres=FALSE)

plot(x)

phyt_raster<-x
crs(phyt_raster) <- "+proj=longlat +datum=WGS84 +no_defs" 

# # the finla data format needs to be like the modelled one: 
# load("/home/ubuntu/Extract_SouthernOcean/Tmp_data/01_extract_modelled_data_SO_means_narmF.RData")
# mean_all[[1]] # raster for size class 1 map
# class      : RasterLayer 
# dimensions : 38, 360, 13680  (nrow, ncol, ncell)
# resolution : 1, 1  (x, y)
# extent     : -180, 180, -78, -40  (xmin, xmax, ymin, ymax)
# crs        : +proj=longlat +datum=WGS84 +no_defs 
# source     : memory
# names      : index_1 
# values     : 1.37047, 7.065759  (min, max)

## phyto and zoo must have the same extent when calculating the ratio
plot(zoop_raster)
bb <- extent(phyt_raster)
zoop_raster <- setExtent(zoop_raster, bb, keepres=FALSE)
plot(zoop_raster)

save(phyt_raster, zoop_raster, file = paste( "/home/ubuntu/Extract_SouthernOcean/Tmp_data/",Sys.Date(),"_03_work_with_empirical.RData", sep = ""))

```
