==> R/04_plot_maps.Rmd <==
---
title: "PLot maps of MEMs and ESMs for the Southern Ocean"
author: "Camilla Novaglio"
date: "01/05/2023"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Set environment 

```{r environment, echo = FALSE}

rm(list=ls())

library(raster) 
library(ggplot2) 
library(dplyr)
library(sf)
library(rnaturalearth)
library(patchwork)
library(colorspace)

select<-dplyr::select
summarise<-dplyr::summarise

my_theme_map<-theme_bw()+ 
  theme(
    text = element_text(size = 9), 
    plot.title = element_text(size = 9),
    axis.title = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank(),
    legend.title=element_text(size = 8), 
    legend.text=element_text(size = 7),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(), 
    panel.background = element_blank(),
    plot.background = element_rect(fill="white"),
    panel.border = element_blank(),
    legend.key.height = unit(0.2, "cm"), # this becomes width if you choose bottom position
    legend.key.width = unit(1, "cm"),
    legend.position= 'bottom' # c(0.5,0.6) # in middle
)

my_theme_map2<-theme_bw()+ 
  theme(
    text = element_text(size = 9), 
    plot.title = element_text(size = 9),
    axis.title = element_blank(),
    axis.text = element_blank(),
    axis.ticks = element_blank(),
    legend.title=element_text(size = 8), 
    legend.text=element_text(size = 7),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(), 
    panel.background = element_blank(),
    plot.background = element_rect(fill="white"),
    panel.border = element_blank(),
    legend.key.height = unit(0.7, "cm"), 
    legend.key.width = unit(0.2, "cm")
  )    

# load functions 
source("/home/ubuntu/Extract_SouthernOcean/R/Plotting_functions_SO.r")
# Checked oob argument in plotting function as well - all good. 

```

## set arguments 

```{r}

# loading data 
removeNA = "narmF" # options narmT and narmF
history_year1 = 1995 # see comments on 01 and 02 re period considered. 
output_date = "2024-01-25" # latest date when the outputs from 01 and 02 were printed 
  
# plankton options (latest agreed version is summer and zoo and Hyear 1995)
season = "summer" # "year" # "summer"
zoopl = "zoo" # "zoo" # "meso"

# trophic amplification 
ltl_type = "sum" # options sum, phyto and zoop. Philip: sum preferred. 

# mapping color 
color_scale = "viridis" # options viridis and monochromatic and divergent (used for trophic amplification)

# mapping number of breaks
breaks_vector_n = 7

# mapping scale 
scale_Fig0 = "raw" # not needed
scale_Fig4 = "log" # new plot for Philip - with log scale and same legend
scale_FigS1 = "raw" # basically same as Fig 5 but with raw values 
scale_FigS2 = "log" # new plot for Philip - with log scale to go with Fig4 and same legend
scale_FigS3 = "raw" # new plot for Philip - scale needs to be raw 
scale_FigS4 = "raw" # new plot for Philip - scale needs to be raw 

# mapping limits
min_Fig0 = 0
max_Fig0_phyt = 8.7
max_Fig0_zoop = 25

min_Fig4 = 0.5 # if up to each dataset limits: e.g. min = phyt_raster@data@min
max_Fig4 = 5 # max = min = phyt_raster@data@max

# S1 has changing max and min vals according to data

min_FigS2 = 0.5
max_FigS2 = 5

min_FigS3 = 0.5 # 0.5 means that the higher trophic level is 1/2 the lower trophic level.
max_FigS3 = 1.5 # 2 means that the higher trophic level is double the lower trophic level. 

min_FigS4 = 0.5 
max_FigS4 = 1.5 

```

## Load data means 

```{r}

# phyto and zoo are different when doing na.rm = T or F because the scale considered changes (i.e. smallest and largest values are removed when na.rm = F for zoo in particular). range of values are smaller and closer together.

# if(removeNA == "narmT"){
#   load("/home/ubuntu/Extract_SouthernOcean/Tmp_data/01_extract_modelled_data_SO_means.RData")
#   load("/home/ubuntu/Extract_SouthernOcean/Tmp_data/02_extract_input_SO_means.RData")
#   load("/home/ubuntu/Extract_SouthernOcean/Tmp_data/03_work_with_empirical.RData")
# }else if (removeNA == "narmF"){
#   load("/home/ubuntu/Extract_SouthernOcean/Tmp_data/01_extract_modelled_data_SO_means_narmF.RData")
#   load("/home/ubuntu/Extract_SouthernOcean/Tmp_data/02_extract_input_SO_means_narmF.RData")
#   load("/home/ubuntu/Extract_SouthernOcean/Tmp_data/03_work_with_empirical.RData")
# }

# load latest outputs
load(paste("/home/ubuntu/Extract_SouthernOcean/Tmp_data/",output_date, "_01_extract_modelled_data_SO_means_",removeNA, "_Hyear_", history_year1, ".RData", sep = "")) # always yearly res as the only possible
load(paste("/home/ubuntu/Extract_SouthernOcean/Tmp_data/",output_date, "_02_extract_input_SO_means_",removeNA,"_", season, "_zooType_", zoopl, "_Hyear_", history_year1, ".RData", sep = ""))
load(paste("/home/ubuntu/Extract_SouthernOcean/Tmp_data/",output_date, "_03_work_with_empirical.RData", sep = ""))  

```

## calculate trophic amplification

Philip: Could you plot the amplification effect for the model. So for each pixel upper trophic level biomass / lower trophic level biomass

```{r}

# amplification factor for Fig S3 (zoopl/phitoplankton)

# zoo/phyto empirical 
names(phyt_raster)<-"index_1" 
phyt_raster@data@min
phyt_raster@data@max

names(zoop_raster)<-"index_1" 
zoop_raster@data@min
zoop_raster@data@max 

tl1_empirical<-zoop_raster/phyt_raster
plot(tl1_empirical)

# # CHECK
# trial1<-as.data.frame(zoop_raster , xy=TRUE)
# trial2<-as.data.frame(phyt_raster, xy=TRUE) %>% rename(index_2 = index_1)
# 
# trial3<-trial1 %>% full_join(trial2) %>% mutate(ratio = index_1/index_2)
# trial4<-rasterFromXYZ(trial3[,c("x", "y", "ratio")])
# plot(trial4)
# plot(tl1_empirical)
# 
# trial5<-as.data.frame(tl1_empirical, xy=TRUE)
# trial1[60:70,]
# trial2[60:70,]
# trial5[60:70,]
# 0.04702734/0.7903568 # 0.0595014

# zoo/phyto modelled 
tl1_modelled<-mean_zoo/mean_phyto
plot(tl1_modelled)

# # check 
# trial1<-as.data.frame(mean_zoo , xy=TRUE)
# trial2<-as.data.frame(mean_phyto, xy=TRUE) %>% rename(index_2 = index_1)
# 
# trial3<-trial1 %>% full_join(trial2) %>% mutate(ratio = index_1/index_2)
# trial4<-rasterFromXYZ(trial3[,c("x", "y", "ratio")])
# plot(trial4)
# plot(tl1_modelled)
# 
# trial5<-as.data.frame(tl1_modelled, xy=TRUE)
# head(trial1)
# head(trial2)
# head(trial5)

```

## Load land map 

```{r}

wmap<-ne_download(scale = 'medium', type = 'map_units')
# crop before converting to circumpolar (as in helper_function_OS for the data)  
bb <- as(extent(-180, 180, -90, -40), 'SpatialPolygons') 
crs(bb) <- crs(wmap)
wmap_SO<-crop(wmap,bb)
# plot(wmap_SO)
# convert to circumpolar
circumpolarCRS <- CRS("+proj=laea +lat_0=-90 +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0")
wmap_SO_circumpolar <- spTransform(wmap_SO, CRS("+proj=laea +lat_0=-90 +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0")) 
# plot(wmap_SO_circumpolar)
wmap_SO_df_circumpolar <- fortify(wmap_SO_circumpolar)
# rename to follow map code below 
antmap_df<-wmap_SO_df_circumpolar

```

## Produce maps of data  

```{r}

### Empirical data check ----

# check patterns are as per Yang et al. 2022 

if(scale_Fig0 == "raw"){
  # legend_name = expression(g~C~m^{-2})
  legend_name = expression(Carbon~biomass~"("~g~C~m^{-2}~")")
}else if(scale_Fig0 == "log"){
  # legend_name = expression(log~g~C~m^{-2})
  legend_name = expression(Carbon~biomass~"("~log~g~C~m^{-2}~")")
}

check_phyto<-plot_SO(wmap_df = antmap_df, data_to_plot = phyt_raster, proj = circumpolarCRS, legend_name = legend_name, min = min_Fig0, max = max_Fig0_phyt, data_type = "phyto", color_scale = "viridis", theme_map = "my_theme_map", scale = scale_Fig0, breaks_vector_n = breaks_vector_n)
# check_phyto<-check_phyto+ggtitle("a) Phytoplankton empirical")
check_phyto<-check_phyto+ggtitle("a) Remotely sensed phytoplankton")

check_zoo<-plot_SO(wmap_df = antmap_df, data_to_plot = zoop_raster, proj = circumpolarCRS, legend_name = legend_name, min = min_Fig0, max = max_Fig0_zoop, data_type = "zoo", color_scale = "viridis", theme_map= "my_theme_map", scale = scale_Fig0, breaks_vector_n = breaks_vector_n)

# check_zoo<-check_zoo+ggtitle("b) Zooplankton empirical")
check_zoo<-check_zoo+ggtitle("b) Empirical zooplankton")

### Empirical data for Fig 4 -----
# Fig4. Depth integrated (surface layer) a) phytoplankton and b) mesozooplankton carbon biomass for the summer months (December to March), where phytoplankton biomass is based on multiyear (1998-2020) mean summer Chl a value (ESA OC-CCI data) while mesozooplankton biomass is from the dataset assembled in Yang et al. 2022 and KRILLBASE (Atkinson et al. 2017). Panels a) and b) are adapted from Fig. 2 in Yang et al. 2022 and use the same data. Depth integrated (water column) c) phytoplankton and d) zooplankton carbon concentration projections averaged across the years 1995-2014 and two Earth-system models, IPSL-CM6A-LR and GFDL-ESM4, made available through the ISIMIP repository (https://data.isimip.org/;  https://doi.org/10.48364/ISIMIP.575744.4). The range of values is kept between 0.5 and 5 g C m-2, in log scale, for better comparison across panels and to highlight geographical patterns (see S-Fig. 1 for raw values). For a comparison with carbon biomass projections for higher trophic levels, see S-Figs 2-4. The code used to produce this figure, and S-Figs 1-4 is available at https://github.com/Fish-MIP/Extract_SouthernOcean.  

if(scale_Fig4 == "raw"){
  # legend_name = expression(g~C~m^{-2})
  legend_name = expression(Carbon~biomass~"("~g~C~m^{-2}~")")
}else if(scale_Fig4 == "log"){
  # legend_name = expression(log~g~C~m^{-2})
  legend_name = expression(Carbon~biomass~"("~log~g~C~m^{-2}~")")
}

names(phyt_raster)<-"index_1" 
phyto_emp_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = phyt_raster, proj = circumpolarCRS, legend_name = legend_name, min = min_Fig4, max = max_Fig4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_Fig4, breaks_vector_n = breaks_vector_n)
phyto_emp_plot<-phyto_emp_plot+ggtitle("a Remotely sensed phytoplankton")

phyt_raster # range
# zoop_raster_trial<-zoop_raster
# zoop_raster_trial[is.na(zoop_raster_trial)]<-Inf

names(zoop_raster)<-"index_1" 
zoop_emp_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = zoop_raster, proj = circumpolarCRS, legend_name = legend_name, min = min_Fig4, max = max_Fig4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_Fig4, breaks_vector_n = breaks_vector_n)
zoop_emp_plot<-zoop_emp_plot+ggtitle("b Empirical zooplankton")

zoop_raster# range

### ESMs data for Fig 4 -----

phyto_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_phyto, proj = circumpolarCRS, legend_name = legend_name, min = min_Fig4, max = max_Fig4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_Fig4, breaks_vector_n = breaks_vector_n)
phyto_plot<-phyto_plot+ggtitle("c Modelled phytoplankton")

mean_phyto # range

zoo_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_zoo, proj = circumpolarCRS, legend_name = legend_name, min = min_Fig4, max = max_Fig4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_Fig4, breaks_vector_n = breaks_vector_n) 
zoo_plot<-zoo_plot+ggtitle("d Modelled zooplankton")

mean_zoo # range

### empirical data for Fig S1 -----
# Fig S1. Depth integrated a) phytoplankton and b) mesozooplankton carbon biomass from Yang et al. 2022 and c) phytoplankton and d) zooplankton carbon concentration projections averaged across IPSL-CM6A-LR and GFDL-ESM4 Earth system models. Please refer to Fig. 4 for more details on data. Here, the range of values in the legend spans the minimum and the maximum value of plankton in each panel.

if(scale_FigS1 == "raw"){
  legend_name = expression(g~C~m^{-2})
  # legend_name = expression(Carbon~biomass~"("~g~C~m^{-2}~")")
}else if(scale_FigS1 == "log"){
  legend_name = expression(log~g~C~m^{-2}) 
  # legend_name = expression(Carbon~biomass~"("~log~g~C~m^{-2}~")")
}

phyto_plot_emp2_S1B<-plot_SO(wmap_df = antmap_df, data_to_plot = phyt_raster, proj = circumpolarCRS, legend_name = legend_name, min = phyt_raster@data@min, max = phyt_raster@data@max, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map2", scale = scale_FigS1, breaks_vector_n = breaks_vector_n)
phyto_plot_emp2_S1B<-phyto_plot_emp2_S1B+ggtitle("a) Remotely sensed phytoplankton")

zoo_plot_emp2_S1B<-plot_SO(wmap_df = antmap_df, data_to_plot = zoop_raster, proj = circumpolarCRS, legend_name = legend_name, min = zoop_raster@data@min, max = zoop_raster@data@max, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map2", scale = scale_FigS1, breaks_vector_n = breaks_vector_n) 
zoo_plot_emp2_S1B<-zoo_plot_emp2_S1B+ggtitle("b) Empirical zooplankton")

### ESMs data for Fig S1 -----

phyto_plot2_S1B<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_phyto, proj = circumpolarCRS, legend_name = legend_name, min =mean_phyto@data@min, max = mean_phyto@data@max, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map2", scale = scale_FigS1, breaks_vector_n = breaks_vector_n)
phyto_plot2_S1B<-phyto_plot2_S1B+ggtitle("c) Modelled phytoplankton")

zoo_plot2_S1B<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_zoo, proj = circumpolarCRS, legend_name = legend_name, min =mean_zoo@data@min, max = mean_zoo@data@max, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map2", scale = scale_FigS1, breaks_vector_n = breaks_vector_n) 
zoo_plot2_S1B<-zoo_plot2_S1B+ggtitle("d) Modelled zooplankton")

### MEMs data for Fig S2 -----
# Fig S2. Depth integrated (water column), size-resolved exploitable fish biomass projections averaged across the years 1995-2014 and six FishMIP Marine Ecosystem Models able to produce this output and forced using IPSL-CM6A-LR and GFDL-ESM4 climate data. Size classes include modelled individuals of a) 1-10g, b) 10-100g, c) 100g-1kg, d) 1-10kg, e) 10-100kg. Data available at https://data.isimip.org/, ISIMIP3b simulation round. The range of values is kept between 0.5 and 5 g C m-2, in log scale, for all size classes for better comparison across plots and with maps of plankton (Fig. 4).

if(scale_FigS2 == "raw"){
  legend_name = expression(Carbon~biomass~"("~g~C~m^{-2}~")")
  # legend_name = bquote(atop("Carbon biomass", ~"("~g~C~m^{-2}~")" ))
}else if(scale_FigS2 == "log"){
  legend_name = expression(Carbon~biomass~"("~log~g~C~m^{-2}~")")
  # legend_name = bquote(atop("Carbon biomass", ~"("~log~g~C~m^{-2}~")" ))
}

mean_all[[1]]
# mean_all[[1]]@data@max

size1_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_all[[1]], proj = circumpolarCRS, legend_name = legend_name, min = min_FigS2, max = max_FigS2, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS2, breaks_vector_n = breaks_vector_n)
size1_plot<-size1_plot+ggtitle("a) Consumer Biomass 1-10g")
# plot(mean_all[[1]])

size2_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_all[[2]], proj = circumpolarCRS, legend_name = legend_name, min = min_FigS2, max = max_FigS2, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS2, breaks_vector_n = breaks_vector_n)
size2_plot<-size2_plot+ggtitle("b) Consumer Biomass 10-100g")
# plot(mean_all[[2]])

size3_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_all[[3]], proj = circumpolarCRS, legend_name = legend_name, min = min_FigS2, max = max_FigS2, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS2, breaks_vector_n = breaks_vector_n)
size3_plot<-size3_plot+ggtitle("c) Consumer Biomass 100g-1kg")
# plot(mean_all[[3]])

size4_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_all[[4]], proj = circumpolarCRS, legend_name = legend_name, min = min_FigS2, max = max_FigS2, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS2, breaks_vector_n = breaks_vector_n)
size4_plot<-size4_plot+ggtitle("d) Consumer Biomass 1-10kg")
# plot(mean_all[[4]])

size5_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_all[[5]], proj = circumpolarCRS, legend_name = legend_name, min = min_FigS2, max = max_FigS2, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS2, breaks_vector_n = breaks_vector_n)
size5_plot<-size5_plot+ggtitle("e) Consumer Biomass 10-100kg")
# plot(mean_all[[5]])

mean_all[[5]]
# mean_all[[5]]@data@max

size6_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = mean_all[[6]], proj = circumpolarCRS, legend_name = legend_name, min = min_FigS2, max = max_FigS2, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS2, breaks_vector_n = breaks_vector_n)
size6_plot<-size6_plot+ggtitle("f) Consumer Biomass >100kg")
# plot(mean_all[[6]])

# mean_all[[6]]@data@min
# mean_all[[6]]@data@max

### ESMs and empirical data for FigS3  ----
# FigS3. Trophic amplification factor given by the ratio between carbon biomass projections of increasing trophic levels, if greater than one indicates higher abundance of the higher trophic level. Specifically, ratio between a) empirical estimates of zooplankton and of phytoplankton, b) modelled projections of zooplankton and of phytoplankton (with extreme values for panel a) well above 1.5; max = 27.1). Please refer to Figure 4 for more details on the data. 

legend_name = "Ratio" # ratio where g C m-2 cancelled out
color_scale = "divergent"
breaks_vector_n = 11

# tl1_empirical@data@max

names(tl1_empirical)<-"index_1"
tl1_empirical_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = tl1_empirical, proj = circumpolarCRS, legend_name = legend_name, min = min_FigS3, max = max_FigS3, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS3, breaks_vector_n = breaks_vector_n) 
tl1_empirical_plot<-tl1_empirical_plot+ggtitle("a) Observed biomass zooplankon / phytoplankton")

names(tl1_modelled)<-"index_1"
tl1_modelled_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = tl1_modelled, proj = circumpolarCRS, legend_name = legend_name, min = min_FigS3, max = max_FigS3, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS3, breaks_vector_n = breaks_vector_n) 
tl1_modelled_plot<-tl1_modelled_plot+ggtitle("a) Modelled biomass zooplankon / phytoplankton")

### ESMs and MEMs data for FigS4 ----
# FigS4. Trophic amplification factor given by the ratio between carbon biomass projections of increasing trophic levels, if greater than one indicates higher abundance of the higher trophic level. Specifically, ratio between a) plankton and consumers biomass of 1-10g, b) consumers biomass of 1-10g and 10-100g, c) consumers biomass of 10-100g and 100g-1kg, d) consumers biomass of 100g-1kg and 1-10kg, e) consumers biomass of 1-10kg and 10-100kg. Please refer to Figure 4 and S-Fig. 2 for more details on plankton and consumers biomass projections, respectively. However, note that for plankton projections, all months instead of the summer months were considered to match the resolution of the consumers biomass projections.

## load the yearly data for plankton and calculate amplification factor for Fig S4 as consumer biomass is at annual resolution. (plankton/biomass1 etc..)
season = "year"

load(paste("/home/ubuntu/Extract_SouthernOcean/Tmp_data/",output_date, "_02_extract_input_SO_means_",removeNA,"_", season, "_zooType_", zoopl, "_Hyear_", history_year1, ".RData", sep = ""))

if(ltl_type == "phyto"){
  ltl<-mean_phyto 
}else if(ltl_type == "zoop"){
  ltl<-mean_zoo
}else if(ltl_type == "sum"){
  ltl<-mean_phyto+mean_zoo
}

# # CHECK
# plot(mean_phyto)
# plot(mean_zoo)
# plot(ltl)
# 
# # CHECK 
# trial1<-as.data.frame(rasterToPoints(mean_phyto))
# trial2<-as.data.frame(rasterToPoints(mean_zoo))
# trial3<-as.data.frame(rasterToPoints(ltl))
# head(trial1)
# head(trial2)
# head(trial3)

tl2<-mean_all[[1]]/ltl
tl3<-mean_all[[2]]/mean_all[[1]]
tl4<-mean_all[[3]]/mean_all[[2]]
tl5<-mean_all[[4]]/mean_all[[3]]
tl6<-mean_all[[5]]/mean_all[[4]]
tl7<-mean_all[[6]]/mean_all[[5]]

# # CHECK 
# plot(tl2)
# plot(mean_all[[1]])
# plot(ltl)
# 
# # CHECK
# trial1<-as.data.frame(rasterToPoints(mean_all[[1]]))
# trial2<-as.data.frame(rasterToPoints(ltl))
# trial3<-as.data.frame(rasterToPoints(tl2))
# head(trial1)
# head(trial2)
# head(trial3)

legend_name = "Ratio" # ratio where g C m-2 cancelled out
color_scale = "divergent"
breaks_vector_n = 11

names(tl2)<-"index_1"
tl2_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = tl2, proj = circumpolarCRS, legend_name = legend_name, min = min_FigS4, max = max_FigS4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS4, breaks_vector_n = breaks_vector_n) 

if(ltl_type == "phyto"){
 tl2_plot<-tl2_plot+ggtitle("a) Biomass 1-10g / phytoplankton")
}else if(ltl_type == "zoop"){
  tl2_plot<-tl2_plot+ggtitle("a) Biomass 1-10g / zooplankton")
}else if(ltl_type == "sum"){
  tl2_plot<-tl2_plot+ggtitle("a) Biomass 1-10g / LTL")
}

names(tl3)<-"index_1"
tl3_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = tl3, proj = circumpolarCRS, legend_name = legend_name, min = min_FigS4, max = max_FigS4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS4, breaks_vector_n = breaks_vector_n)
tl3_plot<-tl3_plot+ggtitle("b) Biomass 10-100g / 1-10g")

names(tl4)<-"index_1"
tl4_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = tl4, proj = circumpolarCRS, legend_name = legend_name, min = min_FigS4, max = max_FigS4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS4, breaks_vector_n = breaks_vector_n)
tl4_plot<-tl4_plot+ggtitle("c) Biomass 100g-1kg / 10-100g")

names(tl5)<-"index_1"
tl5_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = tl5, proj = circumpolarCRS, legend_name = legend_name, min = min_FigS4, max = max_FigS4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS4, breaks_vector_n = breaks_vector_n)
tl5_plot<-tl5_plot+ggtitle("d) Biomass 1-10kg / 100g-1kg")

names(tl6)<-"index_1"
tl6_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = tl6, proj = circumpolarCRS, legend_name = legend_name, min = min_FigS4, max = max_FigS4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS4, breaks_vector_n = breaks_vector_n)
tl6_plot<-tl6_plot+ggtitle("e) Biomass 10-100kg / 1-10kg")

names(tl7)<-"index_1"
tl7_plot<-plot_SO(wmap_df = antmap_df, data_to_plot = tl7, proj = circumpolarCRS, legend_name = legend_name, min = min_FigS4, max = max_FigS4, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map", scale = scale_FigS4, breaks_vector_n = breaks_vector_n)
tl7_plot<-tl7_plot+ggtitle("f) Biomass >100kg / 10-100kg")

```

## Save fianl plots 

```{r}

#### Fig 0 check empirical ----

# specify season for each plot 
season = "summer" # no changes here as the only res available for empirical is summer

p_final_Fig0<-
  check_phyto + check_zoo +
  plot_layout(ncol = 2, nrow = 1)

name_Fig0 = paste("/home/ubuntu/Extract_SouthernOcean/Output/", Sys.Date(),"_" ,"Fig0_checkEmpirical","_",scale_Fig0,"_",season,"_zooType_",zoopl,".pdf", sep = "")

pdf(name_Fig0, width = 7, height = 4.4, bg = "transparent")  
p_final_Fig0
dev.off()

#### Fig 4 ----
season = "summer"

layout <- "
AB
CD
"

p_final_Fig4<-
  phyto_emp_plot + zoop_emp_plot +
  phyto_plot + zoo_plot +
  plot_layout(ncol = 2, nrow = 2, guides = 'collect', heights=c(1,1), widths = c(1,1)) & theme(legend.position = 'bottom')

name_Fig4 = paste("/home/ubuntu/Extract_SouthernOcean/Output/", Sys.Date(),"_" ,"Fig4","_",scale_Fig4,"_",removeNA,"_",season,"_zooType_",zoopl,".pdf", sep = "")

pdf(name_Fig4, width = 6, height = 6, bg = "transparent") 
p_final_Fig4
dev.off()

#### Fig S1 ----
season = "summer"

p_final_FigS1B<-
  phyto_plot_emp2_S1B + zoo_plot_emp2_S1B +
  phyto_plot2_S1B + zoo_plot2_S1B +
  plot_layout(ncol = 2, nrow = 2, heights=c(1,1,1,1), widths = c(1,1,1,1)) # guides = 'collect'

### CHANGED TO FIG S2
name_FigS1B = paste("/home/ubuntu/Extract_SouthernOcean/Output/", Sys.Date(),"_" ,"FigS2","_",scale_FigS1,"_",removeNA,"_",season,"_zooType_",zoopl,".pdf", sep = "")

pdf(name_FigS1B, width = 6, height = 6, bg = "transparent")
p_final_FigS1B
dev.off()

#### Fig S2 ----
season = "year" # not needed here as this is the only possible resolution for consumers biomass 

p_final_FigS2<-size1_plot + size2_plot +
  size3_plot + size4_plot +
  size5_plot + # size6_plot +
  plot_layout(ncol = 3, nrow = 2, guides = 'collect') & theme(legend.position = 'bottom')

## CHANGED TO FIG S3
name_FigS2 = paste("/home/ubuntu/Extract_SouthernOcean/Output/", Sys.Date(),"_" ,"FigS3","_",scale_FigS2,"_",removeNA,"_",season,"_zooType_",zoopl,".pdf", sep = "")

pdf(name_FigS2, width = 8, height = 7, bg = "transparent")
p_final_FigS2
dev.off()

#### Fig S3 ---- 
season = "summer"

p_final_FigS3<-
  tl1_empirical_plot + tl1_modelled_plot +
  plot_layout(ncol = 2, nrow = 1, guides = 'collect') & theme(legend.position = 'bottom')

#### CHANGED TO FIG S1
name_FigS3 = paste("/home/ubuntu/Extract_SouthernOcean/Output/", Sys.Date(),"_" ,"FigS1","_",scale_FigS3,"_",removeNA,"_ltl_type_",ltl_type,"_",season,"_zooType_",zoopl,".pdf", sep = "")

pdf(name_FigS3, width = 7, height = 4.4, bg = "transparent")  
p_final_FigS3
dev.off()

#### Fig S4 ----
season = "year" # this was created using yearly data for both inputs and mem

p_final_FigS4<-tl2_plot + tl3_plot +
  tl4_plot + tl5_plot +
  tl6_plot + # tl7_plot +
  plot_layout(ncol = 3, nrow = 2, guides = 'collect') & theme(legend.position = 'bottom')

name_FigS4 = paste("/home/ubuntu/Extract_SouthernOcean/Output/", Sys.Date(),"_" ,"FigS4","_",scale_FigS4,"_",removeNA,"_ltl_type_",ltl_type,"_",season,"_zooType_",zoopl,".pdf", sep = "")

pdf(name_FigS4, width = 8, height = 7, bg = "transparent")
p_final_FigS4
dev.off()

```

## Plankton plots to send for checking 

```{r}

# ### ESMs data for Map of plankton from IPSL and GFDLS as separate for Colleen to check -----
# 
# load("/home/ubuntu/Extract_SouthernOcean/Tmp_data/02_extract_input_SO.RData") # this are "yearly" value of phyto and zoo for the period 1990-1999.  
# 
# phyto_ipsl<-phyto_ant$`ipsl-cm6a-lr_r1i1p1f1_historical_phyc-vint_60arcmin_global_monthly_1850_2014.nc`
# 
# phyto_ipsl_map<-plot_SO(wmap_df = antmap_df, data_to_plot = phyto_ipsl, proj = circumpolarCRS, legend_name = legend_name, min =phyto_ipsl@data@min, max = phyto_ipsl@data@max, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map2", scale = scale_FigS1, breaks_vector_n = breaks_vector_n)
# phyto_ipsl_map<-phyto_ipsl_map+ggtitle("Phytoplankton IPSL")
# 
# phyto_gfdl<-phyto_ant$`gfdl-esm4_r1i1p1f1_historical_phyc-vint_60arcmin_global_monthly_1850_2014.nc`
# 
# phyto_gfdl_map<-plot_SO(wmap_df = antmap_df, data_to_plot = phyto_gfdl, proj = circumpolarCRS, legend_name = legend_name, min =phyto_gfdl@data@min, max = phyto_gfdl@data@max, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map2", scale = scale_FigS1, breaks_vector_n = breaks_vector_n)
# phyto_gfdl_map<-phyto_gfdl_map+ggtitle("Phytoplankton GFDL")
# 
# phyto_gfdl_map+phyto_ipsl_map
# 
# zoo_ipsl<-zoo_ant$`ipsl-cm6a-lr_r1i1p1f1_historical_zooc-vint_60arcmin_global_monthly_1850_2014.nc`
# 
# zoo_ipsl_map<-plot_SO(wmap_df = antmap_df, data_to_plot = zoo_ipsl, proj = circumpolarCRS, legend_name = legend_name, min =zoo_ipsl@data@min, max = zoo_ipsl@data@max, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map2", scale = scale_FigS1, breaks_vector_n = breaks_vector_n)
# zoo_ipsl_map<-zoo_ipsl_map+ggtitle("Zooplankton IPSL")
# 
# zoo_gfdl<-zoo_ant$`gfdl-esm4_r1i1p1f1_historical_zooc-vint_60arcmin_global_monthly_1850_2014.nc`
# 
# zoo_gfdl_map<-plot_SO(wmap_df = antmap_df, data_to_plot = zoo_gfdl, proj = circumpolarCRS, legend_name = legend_name, min =zoo_gfdl@data@min, max = zoo_gfdl@data@max, data_type = "modelled", color_scale = color_scale, theme_map= "my_theme_map2", scale = scale_FigS1, breaks_vector_n = breaks_vector_n)
# zoo_gfdl_map<-zoo_gfdl_map+ggtitle("Zooplankton GFDL")
# 
# zoo_gfdl_map+zoo_ipsl_map


```
